* Re-organized project hierarchy for character scripts * Modified states to better fit the design pattern and be state-agnostic
