+++ Assets/Scripts/Character.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: a841c47dfef5e4e719ad61aada56ac35
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
