+++ Assets/Scripts/Character/State/MoveState.cs
@@ -1,0 +1,51 @@
+using System;
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class MoveState : State
+{
+	private Matrix4x4 isometricMatrix = Matrix4x4.Rotate(Quaternion.Euler(0, 45, 0));
+	public Vector3 movementInput;
+
+	public MoveState(CharacterMovement character) : base(character)
+	{
+	}
+
+	public override void Enter()
+	{
+		Debug.Log("Entering Move State");
+	}
+
+	public override void HandleInput()
+	{
+		movementInput = new Vector3(Input.GetAxisRaw("Horizontal"), 0, Input.GetAxisRaw("Vertical"));
+		if (Input.GetKeyDown(KeyCode.Space))
+		{
+			character.TransitionToState(new DashState(character));
+		}
+
+		if (Input.GetKeyDown(KeyCode.E))
+		{
+			character.TransitionToState(new InteractState(character));
+		}
+
+		if (Input.GetMouseButtonDown(0))
+		{
+			character.TransitionToState(new SpecialState(character));
+		}
+	}
+
+	public override void FixedUpdate()
+	{
+		movementInput = isometricMatrix.MultiplyPoint3x4(movementInput);
+		movementInput.Normalize();
+		character.rb.MovePosition(character.transform.position +
+		                          movementInput * (character.movementSpeed * Time.deltaTime));
+	}
+
+	public override void Exit()
+	{
+		Debug.Log("Exiting Move State");
+	}
+}
