+++ Assets/Scripts/Character/State/State.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+
 public abstract class State
 {
     protected CharacterMovement character;
